app/requirements.txt
streamlit
tensorflow
Pillow
numpy
pandas
opencv-python
